--- block/__init__.py
@@ -1,3 +1,0 @@
-from block.block import Block
-import block
-import block.merkle
