+++ client/__init__.py
@@ -1,5 +1,2 @@
-from block.block import Block
-from block.block import genesis_block
-from client import core
 from client.cli import CLI
