+++ core/blocks/__init__.py
@@ -1,0 +1,2 @@
+import core.blocks
+import core.blocks.merkle
